refactor: [#1491] move functionality from InMemoryTorrentRepository to TorrentRepository
InMemoryTorrentRepository is now a wrapper over TorrentRepository.

It's planned to make the InMemoryTorrentRepository responsible for triggering events.
