+++ packages/torrent-repository/Cargo.toml
@@ -28,4 +28,6 @@ tracing = "0"
 [dev-dependencies]
 async-std = { version = "1", features = ["attributes", "tokio1"] }
 criterion = { version = "0", features = ["async_tokio"] }
+rand = "0"
 rstest = "0"
+torrust-tracker-test-helpers = { version = "3.0.0-develop", path = "../test-helpers" }
